+++ ui/app/components/clients/attribution.js
@@ -46,12 +46,12 @@ export default class Attribution extends Component {
   // truncate data before sending to chart component
   // move truncating to serializer when we add separate request to fetch and export ALL namespace data
   get barChartTotalClients() {
-    return this.args.totalClientsData.slice(0, 10);
+    return this.args.totalClientsData?.slice(0, 10);
   }
 
   get topClientCounts() {
     // get top namespace or auth method
-    return this.args.totalClientsData[0];
+    return this.args.totalClientsData ? this.args.totalClientsData[0] : null;
   }
 
   get attributionBreakdown() {
@@ -65,8 +65,8 @@ export default class Attribution extends Component {
       return {
         description:
           'This data shows the top ten namespaces by client count and can be used to understand where clients are originating. Namespaces are identified by path. To see all namespaces, export this data.',
-        newCopy: `The new clients in the namespace for this ${dateText}. 
-          This aids in understanding which namespaces create and use new clients 
+        newCopy: `The new clients in the namespace for this ${dateText}.
+          This aids in understanding which namespaces create and use new clients
           ${dateText === 'date range' ? ' over time.' : '.'}`,
         totalCopy: `The total clients in the namespace for this ${dateText}. This number is useful for identifying overall usage volume.`,
       };
@@ -74,7 +74,7 @@ export default class Attribution extends Component {
       return {
         description:
           'This data shows the top ten authentication methods by client count within this namespace, and can be used to understand where clients are originating. Authentication methods are organized by path.',
-        newCopy: `The new clients used by the auth method for this ${dateText}. This aids in understanding which auth methods create and use new clients 
+        newCopy: `The new clients used by the auth method for this ${dateText}. This aids in understanding which auth methods create and use new clients
         ${dateText === 'date range' ? ' over time.' : '.'}`,
         totalCopy: `The total clients used by the auth method for this ${dateText}. This number is useful for identifying overall usage volume. `,
       };
