+++ ui/app/controllers/vault/cluster/clients/history.js
@@ -1,0 +1,3 @@
+import Controller from '@ember/controller';
+
+export default class HistoryController extends Controller {}
