+++ ui/app/controllers/vault/cluster/clients/index.js
@@ -1,8 +1,3 @@
 import Controller from '@ember/controller';
 
-export default class ClientsController extends Controller {
-  queryParams = ['tab', 'start', 'end']; // ARG TODO remove
-  tab = null;
-  start = null;
-  end = null;
-}
+export default class ClientsController extends Controller {}
