+++ ui/app/router.js
@@ -18,7 +18,8 @@ Router.map(function () {
       this.mount('open-api-explorer', { path: '/api-explorer' });
       this.route('license');
       this.route('clients', function () {
-        this.route('index', { path: '/' });
+        this.route('history');
+        this.route('config');
         this.route('edit');
       });
       this.route('storage', { path: '/storage/raft' });
