+++ ui/app/routes/vault/cluster/clients/history.js
@@ -2,7 +2,25 @@ import Route from '@ember/routing/route';
 import RSVP from 'rsvp';
 import { action } from '@ember/object';
 
-export default class ClientsRoute extends Route {
+export default class HistoryRoute extends Route {
+  async getActivity(start_time) {
+    try {
+      return this.store.queryRecord('clients/activity', { start_time });
+    } catch (e) {
+      // ARG TODO handle
+      return e;
+    }
+  }
+
+  async getLicense() {
+    try {
+      return this.store.queryRecord('license', {});
+    } catch (e) {
+      // ARG TODO handle
+      return e;
+    }
+  }
+
   async getVersionHistory() {
     try {
       let arrayOfModels = [];
@@ -21,16 +39,27 @@ export default class ClientsRoute extends Route {
     }
   }
 
+  parseRFC3339(timestamp) {
+    // convert '2021-03-21T00:00:00Z' --> ['2021', 2] (e.g. 2021 March, month is zero indexed)
+    return timestamp
+      ? [timestamp.split('-')[0], Number(timestamp.split('-')[1].replace(/^0+/, '')) - 1]
+      : null;
+  }
+
   async model() {
     let config = await this.store.queryRecord('clients/config', {}).catch((e) => {
       console.debug(e);
       // swallowing error so activity can show if no config permissions
       return {};
     });
+    let license = await this.getLicense();
+    let activity = await this.getActivity(license.startTime);
 
     return RSVP.hash({
       config,
-      monthly: await this.store.queryRecord('clients/monthly', {}),
+      activity,
+      startTimeFromLicense: this.parseRFC3339(license.startTime),
+      endTimeFromResponse: activity ? this.parseRFC3339(activity.endTime) : null,
       versionHistory: this.getVersionHistory(),
     });
   }
@@ -38,7 +67,7 @@ export default class ClientsRoute extends Route {
   @action
   async loading(transition) {
     // eslint-disable-next-line ember/no-controller-access-in-routes
-    let controller = this.controllerFor('vault.cluster.clients.index');
+    let controller = this.controllerFor('vault.cluster.clients.history');
     if (controller) {
       controller.currentlyLoading = true;
       transition.promise.finally(function () {
