+++ ui/app/serializers/clients/activity.js
@@ -48,8 +48,8 @@ export default class ActivitySerializer extends ApplicationSerializer {
     return object;
   }
 
-  rfc33395ToMonthYear(timestamp) {
-    // return ['2021', 2] (e.g. 2021 March, make 0-indexed)
+  parseRFC3339(timestamp) {
+    // convert '2021-03-21T00:00:00Z' --> ['2021', 2] (e.g. 2021 March, month is zero indexed)
     return timestamp
       ? [timestamp.split('-')[0], Number(timestamp.split('-')[1].replace(/^0+/, '')) - 1]
       : null;
@@ -65,8 +65,8 @@ export default class ActivitySerializer extends ApplicationSerializer {
       response_timestamp,
       by_namespace: this.flattenDataset(payload.data.by_namespace),
       total: this.homogenizeClientNaming(payload.data.total),
-      formatted_end_time: this.rfc33395ToMonthYear(payload.data.end_time),
-      formatted_start_time: this.rfc33395ToMonthYear(payload.data.start_time),
+      formatted_end_time: this.parseRFC3339(payload.data.end_time),
+      formatted_start_time: this.parseRFC3339(payload.data.start_time),
     };
     delete payload.data.by_namespace;
     delete payload.data.total;
@@ -81,7 +81,6 @@ payload.data.by_namespace = [
   {
     namespace_id: '5SWT8',
     namespace_path: 'namespacelonglonglong4/',
-    _comment1: 'client counts are nested within own object', 
     counts: {
       entity_clients: 171,
       non_entity_clients: 20,
@@ -103,7 +102,6 @@ payload.data.by_namespace = [
 transformedPayload.by_namespace = [
   {
     label: 'namespacelonglonglong4/',
-    _comment2: 'remove nested object', 
     entity_clients: 171,
     non_entity_clients: 20,
     clients: 191,
