+++ ui/app/templates/components/clients/horizontal-bar-chart.hbs
@@ -1,4 +1,5 @@
 <svg
+  data-test-horizontal-bar-chart
   class="chart is-horizontal"
   {{on "mouseleave" this.removeTooltip}}
   {{did-insert this.renderChart @dataset}}
