+++ ui/app/templates/vault/cluster/clients/edit.hbs
@@ -6,4 +6,4 @@
   </p.levelLeft>
 </PageHeader>
 
-<Clients::Config @model={{this.model}} @mode="edit" />
+<Clients::Config @model={{@model}} @mode="edit" />
