+++ ui/app/templates/vault/cluster/clients/history.hbs
@@ -1,0 +1,3 @@
+<Clients::Dashboard @model={{@model}} />
+
+<Clients::History @model={{@model}} @isLoading={{this.currentlyLoading}} />
