+++ ui/app/templates/vault/cluster/clients/index.hbs
@@ -1,69 +1,2 @@
-<PageHeader as |p|>
-  <p.levelLeft>
-    <h1 class="title is-3">
-      Vault Client Count
-    </h1>
-  </p.levelLeft>
-</PageHeader>
-
-<div class="tabs-container box is-bottomless is-marginless is-fullwidth is-paddingless" data-test-pricing-metrics>
-  {{! template-lint-configure no-unknown-arguments-for-builtin-components "warn" }}
-  <nav class="tabs">
-    <ul>
-      {{! template-lint-configure no-unknown-arguments-for-builtin-components "warn" }}
-      <LinkTo
-        @route="vault.cluster.clients.index"
-        @query={{hash tab="current" start=null end=null}}
-        @tagName="li"
-        @activeClass="is-active"
-      >
-        <LinkTo
-          @route="vault.cluster.clients.index"
-          @query={{hash tab="current" start=null end=null}}
-          data-test-usage-tab={{true}}
-        >
-          Current month
-        </LinkTo>
-      </LinkTo>
-      <LinkTo @route="vault.cluster.clients.index" @query={{hash tab="dashboard"}} @tagName="li" @activeClass="is-active">
-        <LinkTo @route="vault.cluster.clients.index" @query={{hash tab="dashboard"}} data-test-usage-tab={{true}}>
-          History
-        </LinkTo>
-      </LinkTo>
-      {{#if this.model.config.configPath.canRead}}
-        <LinkTo
-          @route="vault.cluster.clients.index"
-          @query={{hash tab="config" start=null end=null}}
-          @tagName="li"
-          @activeClass="is-active"
-        >
-          <LinkTo
-            @route="vault.cluster.clients.index"
-            @query={{hash tab="config" start=null end=null}}
-            data-test-configuration-tab={{true}}
-          >
-            Configuration
-          </LinkTo>
-        </LinkTo>
-      {{/if}}
-      {{! template-lint-configure no-unknown-arguments-for-builtin-components "on" }}
-    </ul>
-  </nav>
-</div>
-
-{{#if (eq this.tab "config")}}
-  <Toolbar>
-    <ToolbarActions>
-      {{#if this.model.config.configPath.canUpdate}}
-        <LinkTo @route="vault.cluster.clients.edit" class="toolbar-link">
-          Edit configuration
-        </LinkTo>
-      {{/if}}
-    </ToolbarActions>
-  </Toolbar>
-  <Clients::Config @model={{this.model.config}} @isLoading={{this.currentlyLoading}} />
-{{else if (eq this.tab "current")}}
-  <Clients::Current @tab={{this.tab}} @model={{this.model}} @isLoading={{this.currentlyLoading}} />
-{{else if (eq this.tab "dashboard")}}
-  <Clients::Dashboard @tab={{this.tab}} @model={{this.model}} @isLoading={{this.currentlyLoading}} />
-{{/if}}
+<Clients::Dashboard @model={{@model}} />
+<Clients::Current @model={{@model}} @isLoading={{this.currentlyLoading}} />
